ProxyMatcher: return self in unwrap() if proxy is not armed yet
It can happen, with certain grammars such as the one provided in issue #1, that
a ProxyMatcher is a proxy for another, "unarmed" proxy; in this case the target
is null, and when attempting to set the label, a NullPointerException is thrown
since .unwrap() systematically returns the target.

Change it so that it return itself instead if there is no target.

See https://github.com/parboiled1/parboiled-core/issues/1.