Issue #1 redux: same fix, different name
The SSCCE shows that the NPE appears on basically infinitely recursive
grammars... This patch therefore only delays the problem.

Rename .unwrap() to .tryUnwrap() nevertheless and reapply fix as is; but
detecting the problem before that would be even better. After all, it occurs
at parser build time, might as well avoid the user to shoot itself in the foot!